clean up vehicle detail and move vehicle assign component
